Refactor navigation and update game saves handling in AuthController
