+++ private/views/templates/navbar.php
@@ -2,25 +2,45 @@
 global $theme, $changelog;
 global $require;
 
-$lastVisitedSaveGame = '';
-if (isset($_SESSION['userId'])){
-    if ($_SESSION['userId'] != '' && $_SESSION['userId'] != null){
-        $lastVisitedSaveGame = GameSaves::getLastVisitedSaveGame() ?? GameSaves::getSaveGamesByUser($_SESSION['userId'])[0]->id ?? null;
-        if ($require === '/game_save') {
-            $lastVisitedSaveGame = $_GET['id'];
-        }
+class NavItem {
+    public string $url;
+    public string $label;
+    public bool $active;
+
+    public function __construct(string $url, string $label, bool $active = false) {
+        $this->url = $url;
+        $this->label = $label;
+        $this->active = $active;
+    }
+}
+
+class DropdownNavItem {
+    public string $url;
+    public string $label;
+    public bool $active;
+    public array $items; // Array of navItem objects
+
+    public function __construct(string $url, string $label, array $items, bool $active = false) {
+        $this->url = $url;
+        $this->label = $label;
+        $this->active = $active;
+        $this->items = $items;
     }
 }
 
 $navItems = [
-    '/home' => 'Home',
+        new NavItem('/home', 'Home', $require === '/home'),
 ];
 
-if ($lastVisitedSaveGame != '') {
-    $navItems['/game_save?id=' . $lastVisitedSaveGame] = 'Game Save';
+if (isset($_SESSION['userId'])) {
+    $saveGamesDropdownItems = [];
+    foreach (GameSaves::getSaveGamesByUser($_SESSION['userId']) as $saveGame) {
+        $saveGamesDropdownItems[] = new NavItem('/game_save?id=' . $saveGame->id, htmlspecialchars($saveGame->title), $require === '/game_save' && $_GET['id'] == $saveGame->id);
+    }
+    $navItems[] = new DropdownNavItem('/game_saves', 'Save Games', $saveGamesDropdownItems, $require === '/game_saves');
 }
 
-$navItems['/helpfulLinks'] = 'Helpful Links';
+$navItems[] = new NavItem('/helpfulLinks', 'Helpful Links', $require === '/helpfulLinks');
 
 
 ?>
@@ -49,15 +69,57 @@ $navItems['/helpfulLinks'] = 'Helpful Links';
         </button>
         <div class="collapse navbar-collapse" id="navbarSupportedContent">
             <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-                <?php foreach ($navItems as $url => $label) :?>
-
-                    <li class="nav-item">
-                        <a class="nav-link <?php echo ($require === explode('?', $url)[0]
-                        ) ? 'active' : ''; ?>" aria-current="page" href="<?= $url ?>">
-                            <?= $label ?>
-                        </a>
-                    </li>
+                <?php foreach ($navItems as $navItem): ?>
+                    <?php if (!empty($navItem->items)) : ?>
+                        <!-- Dropdown logic -->
+                        <li class="nav-item dropdown">
+                            <?php if (!empty($navItem->url) && $navItem->url !== '#') : ?>
+                                <!-- Split button: main link + dropdown -->
+                                <div class="btn-group">
+                                    <a href="<?= $navItem->url ?>"
+                                       class="nav-link btn btn-link me-2 me-lg-0 <?= $navItem->active ? 'active' : '' ?>">
+                                        <?= $navItem->label ?>
+                                    </a>
+                                    <button type="button"
+                                            class="btn btn-link nav-link dropdown-toggle dropdown-toggle-split <?= $navItem->active ? 'active' : '' ?>"
+                                            data-bs-toggle="dropdown" aria-expanded="false">
+                                        <span class="visually-hidden">Toggle Dropdown</span>
+                                    </button>
+                                    <ul class="dropdown-menu">
+                                        <?php foreach ($navItem->items as $item): ?>
+                                            <li>
+                                                <a class="dropdown-item <?= $item->active ? 'active' : '' ?>"
+                                                   href="<?= $item->url ?>"><?= $item->label ?></a>
+                                            </li>
+                                        <?php endforeach; ?>
+                                    </ul>
+                                </div>
+                            <?php else : ?>
+                                <!-- Normal dropdown toggle -->
+                                <a class="nav-link dropdown-toggle <?= $navItem->active ? 'active' : '' ?>"
+                                   href="#" id="navbarDropdown" role="button"
+                                   data-bs-toggle="dropdown" aria-expanded="false">
+                                    <?= $navItem->label ?>
+                                </a>
+                                <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
+                                    <?php foreach ($navItem->items as $item): ?>
+                                        <li>
+                                            <a class="dropdown-item <?= $item->active ? 'active' : '' ?>"
+                                               href="<?= $item->url ?>"><?= $item->label ?></a>
+                                        </li>
+                                    <?php endforeach; ?>
+                                </ul>
+                            <?php endif; ?>
+                        </li>
+                    <?php else: ?>
+                        <!-- Regular nav item -->
+                        <li class="nav-item me-lg-2 mt-lg-0 mt-2">
+                            <a class="nav-link <?= $navItem->active ? 'active' : '' ?>"
+                               href="<?= $navItem->url ?>"><?= $navItem->label ?></a>
+                        </li>
+                    <?php endif; ?>
                 <?php endforeach; ?>
+
             </ul>
             <div class="d-flex">
                 <ul class="navbar-nav me-auto mb-2 mb-lg-0">
@@ -117,3 +179,20 @@ $navItems['/helpfulLinks'] = 'Helpful Links';
         }
     });
 </script>
+<!--<script>-->
+<!--    document.addEventListener('DOMContentLoaded', function () {-->
+<!--        const dropdownLink = document.getElementById('navbarDropdown');-->
+<!--        dropdownLink.addEventListener('click', function (e) {-->
+<!--            e.preventDefault()-->
+<!--            // Go directly to href instead of toggling the dropdown-->
+<!--            window.location.href = dropdownLink.getAttribute('href');-->
+<!--        });-->
+<!--    });-->
+<!--</script>-->
+<!--<style>-->
+<!--    /* Show dropdown on hover */-->
+<!--    .nav-item.dropdown:hover .dropdown-menu {-->
+<!--        display: block;-->
+<!--        margin-top: 0;-->
+<!--    }-->
+<!--</style>-->
